Isaac movement

+++ src/main/java/com/game/void_seekers/character/base/PlayableCharacter.java
@@ -59,7 +59,7 @@ public abstract class PlayableCharacter extends GameCharacter implements Attack 
         setLuck(luck);
     }
 
-    public double getSpeed() {
+    public int getSpeed() {
         return speed;
     }
